mysite/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.2
Django==4.0.4
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-extensions==3.1.3
django-redis==5.2.0
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==20.0.1
pyparsing==3.0.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
Werkzeug==2.0.1
wrapt==1.13.3
